requirements.txt
kafka-python==2.0.2
git+https://github.com/obervinov/logger-package.git#egg=logger